+++ spec/moped/query_spec.rb
@@ -375,8 +375,7 @@ describe Moped::Query do
           operation.selector.should eq(
             "$query" => { likes: { "$exists" => false }},
             "$explain" => true,
-            "$orderby" => { _id: 1 },
-            "$limit" => -1
+            "$orderby" => { _id: 1 }
           )
         end
 
@@ -412,8 +411,7 @@ describe Moped::Query do
         it "updates to a mongo advanced selector" do
           operation.selector.should eq(
             "$query" => { created_at: { "$exists" => false }},
-            "$explain" => true,
-            "$limit" => -1
+            "$explain" => true
           )
         end
 
@@ -450,8 +448,7 @@ describe Moped::Query do
           operation.selector.should eq(
             "$query" => { likes: { "$exists" => false }},
             "$explain" => true,
-            "$hint" => { _id: 1 },
-            "$limit" => -1
+            "$hint" => { _id: 1 }
           )
         end
 
@@ -461,6 +458,34 @@ describe Moped::Query do
 
         it "scans more than one object" do
           explain["nscannedObjects"].should eq(2)
+        end
+      end
+
+      context "when a limit exists" do
+
+        before do
+          4.times do |n|
+            users.insert({ likes: n })
+          end
+        end
+
+        let(:explain) do
+          users.find(likes: { "$gt" => 1 }).limit(2).explain
+        end
+
+        let(:stats) do
+          Support::Stats.collect { explain }
+        end
+
+        let(:operation) do
+          stats[node_for_reads].grep(Moped::Protocol::Query).last
+        end
+
+        it "updates to a mongo advanced selector" do
+          operation.selector.should eq(
+            "$query" => { likes: { "$gt" => 1 }},
+            "$explain" => true
+          )
         end
       end
     end